requirements.txt
# Run `pip install -r requirements.txt` to install these dependencies
# Activate the virtual environment with `source venv/bin/activate`
beautifulsoup4==4.12.3
jinja2==3.1.4
markdown==3.7
pillow==11.0.0
pyYAML==6.0.2
typogrify==2.0.7
watchdog==6.0.0